+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 3.0.1-RC2
 * deleted `ComponentTestCase` and `HelperTestCase` classes;
 * deleted `BreadcrumbsHelper` class;
+* deleted `getMockForHelper()`, `getTableClassNameFromAlias()` and `getPluginName()` methods provided by the `MockTrait`;
 * deleted `is('ip')`, `is('matchingIp')`, `is('localhost')`, `is('prefix')` and `is('url')` request detector methods.
 
 ### 3.0.0-RC1